dom/webgpu/mochitest/test_buffer_mapping.html

<!DOCTYPE HTML>
<html>
<head>
<meta charset='utf-8'>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body>
<script>

ok(SpecialPowers.getBoolPref('dom.webgpu.enabled'), 'Pref should be enabled.');

const func = async function() {
  const adapter = await navigator.gpu.requestAdapter();
  const device = await adapter.requestDevice();
  const [buffer, mapping] = await device.createBufferMapped({
    size: 4,
    usage: GPUBufferUsage.MAP_READ | GPUBufferUsage.COPY_SRC, //TODO: remove the dummy usage
  });
  new Float32Array(mapping).set([1.0]);
  buffer.unmap();
  const data = await buffer.mapReadAsync();
  const value = new Float32Array(data)[0];
  buffer.unmap();
  ok(value == 1.0, 'value == 1.0');
};

SimpleTest.waitForExplicitFinish();
func().finally(() => SimpleTest.finish());

</script>
</body>
</html>
